+++ rector.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\CodingStyle\Rector\Catch_\CatchExceptionNameMatchingTypeRector;
+use Rector\CodingStyle\Rector\PostInc\PostIncDecToPreIncDecRector;
+use Rector\CodingStyle\Rector\String_\SymplifyQuoteEscapeRector;
+use Rector\Config\RectorConfig;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\YieldDataProviderRector;
+
+return RectorConfig::configure()
+    ->withPaths([
+        __DIR__ . '/examples',
+        __DIR__ . '/src',
+        __DIR__ . '/tests',
+    ])
+    ->withPhpSets()
+    ->withPreparedSets(
+        deadCode: true,
+        codeQuality: true,
+        codingStyle: true,
+        typeDeclarations: true,
+        privatization: true,
+        instanceOf: true,
+        earlyReturn: true,
+        strictBooleans: true,
+        rectorPreset: true,
+        phpunitCodeQuality: true,
+    )
+    ->withSkip([
+        CatchExceptionNameMatchingTypeRector::class,
+        PostIncDecToPreIncDecRector::class,
+        SymplifyQuoteEscapeRector::class,
+        YieldDataProviderRector::class,
+    ])
+;