+++ src/Psr/HttpClient.php
@@ -72,7 +72,6 @@ final class HttpClient implements RequestFactoryInterface, StreamFactoryInterfac
      * The stream SHOULD be created with a temporary resource.
      *
      * @param string $content String content with which to populate the stream.
-     * @return StreamInterface
      */
     public function createStream(string $content = ''): StreamInterface
     {
@@ -88,7 +87,7 @@ final class HttpClient implements RequestFactoryInterface, StreamFactoryInterfac
     {
         $headers = [];
 
-        foreach ($request->getHeaders() as $key => $header) {
+        foreach (array_keys($request->getHeaders()) as $key) {
             $headers[$key] = $request->getHeaderLine($key);
         }
 
@@ -120,7 +119,6 @@ final class HttpClient implements RequestFactoryInterface, StreamFactoryInterfac
      * @param string $filename Filename or stream URI to use as basis of stream.
      * @param string $mode Mode with which to open the underlying filename/stream.
      *
-     * @return StreamInterface
      * @throws \RuntimeException If the file cannot be opened.
      * @throws \InvalidArgumentException If the mode is invalid.
      */
@@ -135,8 +133,6 @@ final class HttpClient implements RequestFactoryInterface, StreamFactoryInterfac
      * The stream MUST be readable and may be writable.
      *
      * @param resource $resource PHP resource to use as basis of stream.
-     *
-     * @return StreamInterface
      */
     public function createStreamFromResource($resource): StreamInterface
     {