+++ tests/Psr/Response/GetProtocolVersionTest.php
@@ -21,6 +21,7 @@ final class GetProtocolVersionTest extends TestCase
         $requestsResponse = new RequestsResponse();
         $requestsResponse->status_code = 200;
         $requestsResponse->protocol_version = 1.0;
+
         $response = Response::fromResponse($requestsResponse);
 
         TestCase::assertSame('1.0', $response->getProtocolVersion());
@@ -36,6 +37,7 @@ final class GetProtocolVersionTest extends TestCase
         $requestsResponse = new RequestsResponse();
         $requestsResponse->status_code = 200;
         $requestsResponse->protocol_version = false;
+
         $response = Response::fromResponse($requestsResponse);
 
         TestCase::assertSame('1.1', $response->getProtocolVersion());