+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Run PHPStan with activated strict rules.
 - Run tests with PHPUnit 11 and 12.
+- Add `rector.php` config file to run [rector](https://getrector.com) for code improvements.
 
 ### Changed